refactor: P1-T3 - 拆分ImmersiveNotationOverlayController为Service与View
